refactor: Improve profile component architecture and remove unnecessary props
- Remove unused onboardingData prop from Profile component
- Remove redundant 'No user state' check (handled by middleware)
- Remove unnecessary onRetakeOnboarding prop passing
- Move retake onboarding logic directly into FitnessProfileCard
- Add loading.tsx file convention for better UX
- Clean up component interfaces and reduce coupling
- Improve separation of concerns and maintainability
